# perf(ext/web): optimize transferArrayBuffer