--- a/ext/web/06_streams.js
+++ b/ext/web/06_streams.js
@@ -220,7 +220,8 @@
    */
   function transferArrayBuffer(O) {
     const v = ops.op_transfer_arraybuffer(O);
-    v[isFakeDetached] = true;
+    O[isFakeDetached] = true;
+    return v;
   }
 
   /**