@bartlomieju 2022-10-25
https://github.com/denoland/deno/pull/16307 has landed, please rebase this PR @littledivy